+++ src/data_source_manager.cpp
@@ -1,0 +1,192 @@
+#include "data_source_manager.hpp"
+
+
+DataSourceManager *DataSourceManager::instance = 0;
+
+
+DataSourceManager::DataSourceManager()
+{
+}
+
+
+DataSourceManager::~DataSourceManager()
+{
+    removeAllSources(false);
+}
+
+
+DataSeriesPointer DataSourceManager::findSeries(QString source_label, QString series_label)
+{
+    auto source = getSourceByLabel(source_label);
+
+    if (source.isNull()) return DataSeriesPointer(nullptr);
+
+    return source->getSeriesByLabel(series_label);
+}
+
+
+QStringList DataSourceManager::getSourceLabels() const
+{
+    QStringList labels;
+
+    for (auto src : sources)
+    {
+        if (src.isNull()) continue;
+
+        labels.push_back(src->getLabel());
+    }
+
+    return labels;
+}
+
+
+DataSourcePointer DataSourceManager::getSourceByIndex(unsigned int idx)
+{
+    if (idx < sources.size()) return sources.at(idx);
+
+    return DataSourcePointer(nullptr);
+}
+
+
+DataSourcePointer DataSourceManager::getSourceByLabel(QString label)
+{
+    for (auto src : sources)
+    {
+        if (src.isNull()) continue;
+
+        if (src->getLabel() == label)
+        {
+            return src;
+        }
+    }
+
+    return DataSourcePointer(nullptr);
+}
+
+
+bool DataSourceManager::addSource(DataSourcePointer source)
+{
+    if (source.isNull()) return false;
+
+    for (auto src : sources)
+    {
+        if (src == source)
+        {
+            qInfo() << "Ignoring duplicate source:" << source->getLabel();
+
+            return false;
+        }
+    }
+
+    sources.push_back(source);
+
+    connect(source.data(), &DataSource::dataChanged, this, &DataSourceManager::onDataChanged);
+
+    emit sourcesChanged();
+
+    return true;
+}
+
+
+/**
+ * @brief DataSourceManager::addSource adds a DataSource with the given title
+ * @param label is the label of the DataSource to add
+ * @return true if the source was added, false if a DataSource with the given title already existed
+ */
+bool DataSourceManager::addSource(QString label, QString description)
+{
+    if (getSourceByLabel(label).isNull())
+    {
+        return addSource(new DataSource(label, description));
+    }
+
+    // Source with provided label already exists!
+    return false;
+}
+
+
+/**
+ * @brief DataSourceManager::removeSource removes the DataSource from this DataSourceManager
+ * @param source is a shared pointer to the DataSource
+ * @return true if the source was removed, else false
+ */
+bool DataSourceManager::removeSource(DataSourcePointer source)
+{
+    for (auto idx = 0; idx < sources.size(); idx++)
+    {
+        auto src = sources.at(idx);
+
+        if (src == source)
+        {
+            sources.removeAt(idx);
+            emit sourcesChanged();
+            return true;
+        }
+    }
+
+    return false;
+}
+
+
+/**
+ * @brief DataSourceManager::removeSourceByIndex removes the DataSource at the specified index
+ * @param idx - Index of the DataSource to remove
+ * @param update - If true, emit a "sourcesChanged" event
+ * @return true if the source was removed, else false (in the case of an invalid index)
+ */
+bool DataSourceManager::removeSourceByIndex(unsigned int idx, bool update)
+{
+    if (idx < sources.size())
+    {
+        sources.removeAt(idx);
+
+        if (update)
+        {
+            emit sourcesChanged();
+        }
+
+        return true;
+    }
+
+    return false;
+}
+
+
+/**
+ * @brief DataSourceManager::removeSourceByLabel - Remove the first DataSource which matches the specified label
+ * @param label - QString label
+ * @param update - If true, emit a "sourcesChanged" event
+ * @return true if a source matching the provided label was found (and subsequentyly deleted)
+ */
+bool DataSourceManager::removeSourceByLabel(QString label, bool update)
+{
+    for (int idx = 0; idx < sources.size(); idx++)
+    {
+        auto src = sources.at(idx);
+
+        if (!src.isNull() && src->getLabel() == label)
+        {
+            return removeSourceByIndex(idx, update);
+        }
+    }
+
+    return false;
+}
+
+
+/**
+ * @brief DataSourceManager::removeAllSources removes all DataSource objects from this DataSourceManager
+ * @param update - if true, emit a "sourcesChanged" event
+ */
+void DataSourceManager::removeAllSources(bool update)
+{
+    while (sources.count() > 0)
+    {
+        removeSourceByIndex(0, update);
+    }
+
+    if (update)
+    {
+        emit sourcesChanged();
+    }
+}
